+++ prg/__init__.py
@@ -9,10 +9,10 @@ Example
 -------
 >>> import prg
 >>> prg.__version__
-'0.3.3'
+'0.3.4'
 """
 
-__version__ = "0.3.3"
+__version__ = "0.3.4"
 
 __all__ = [
     "__version__",