+++ prg/copulas/archimedean/__init__.py
@@ -6,10 +6,12 @@ from prg.copulas.archimedean.a14       import CopulaA14
 from prg.copulas.archimedean.joe       import CopulaJoe
 from prg.copulas.archimedean.survival  import SurvivalCopula, SurvivalClayton, SurvivalGH, SurvivalJoe
 from prg.copulas.archimedean.bb1       import CopulaBB1
+from prg.copulas.archimedean.amh       import CopulaAMH
 
 __all__ = [
     'CopulaGH', 'CopulaClayton', 'CopulaFrank', 'CopulaA12', 'CopulaA14',
     'CopulaJoe',
     'SurvivalCopula', 'SurvivalClayton', 'SurvivalGH', 'SurvivalJoe',
     'CopulaBB1',
+    'CopulaAMH',
 ]