+++ prg/copulas/explicit/__init__.py
@@ -1,5 +1,6 @@
 from prg.copulas.explicit.product       import CopulaProduct
 from prg.copulas.explicit.fgm           import CopulaFGM
 from prg.copulas.explicit.cubic_section import CopulaCubSec
+from prg.copulas.explicit.plackett      import CopulaPlackett
 
-__all__ = ['CopulaProduct', 'CopulaFGM', 'CopulaCubSec']
+__all__ = ['CopulaProduct', 'CopulaFGM', 'CopulaCubSec', 'CopulaPlackett']