+++ prg/tests/test_smoke.py
@@ -5,4 +5,4 @@ import prg
 
 def test_version_is_defined():
     assert isinstance(prg.__version__, str)
-    assert prg.__version__ == "0.3.3"
+    assert prg.__version__ == "0.3.4"